# perf: only allocate vector for `inner_get_car_positions` when necessary