`inner_get_car_positions` is a hot function, and it recurses.  In its hottest invocation (including it's own recursion) we only care about the position of the last car, so we can avoid allocating all the intermediate positions. This complicates the method a a bit, but it's a pretty hefty performance gain: 12-13% less runtime for me.

Command ran:

    $ cargo flamegraph --bin run_scenario -- --disable_turn_conflicts --infinite_parking --disable_block_the_box --alerts=silence --disable_recalc_lc data/system/scenarios/downtown/weekday.bin --hours=8

**before**
```
- After 3m26.7s, the sim is at 08:00:00.0. 19,606 live agents
Advance sim to 08:00:00.0 took 206.7296s
run simulation... plus 0.0000s
run simulation took 206.7296s
```

<img width="1792" alt="Screen Shot 2020-10-19 at 12 29 39 PM" src="https://user-images.githubusercontent.com/217057/96502533-c8a08980-1206-11eb-92f1-8336c8809f41.png">
(search highlight is for "malloc" which accounts for most of the improvement)

**after**

```
- After 2m59.9s, the sim is at  07:59:53.4. 18,968 live agents
Advance sim to 08:00:00.0 took 180.3246s                             
run simulation... plus 0.0000s                                             
run simulation took 180.3246s
```

<img width="1792" alt="Screen Shot 2020-10-19 at 12 29 21 PM" src="https://user-images.githubusercontent.com/217057/96502544-cccca700-1206-11eb-83fa-9bc2999892b9.png">