@dabreegster 2020-10-19
Awesome, thank you! I saw `get_car_positions` was a clear hot-spot, but thought that's fundamental to what it does, and didn't dig in and see the malloc. I don't think the new code is much more complicated.